requirements.txt
# Core Dependencies
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Social Media APIs
atproto==0.0.46

# Google Cloud
google-cloud-language==2.13.0
google-cloud-storage==2.14.0
functions-framework==3.5.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Dashboard
streamlit==1.29.0
plotly==5.18.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
